Fix a couple of issues in the ab resource class.
THIS IS BINARY INCOMPATIBLE!

Note to translators: It also adds two new i18n strings (sorry). The strings
are identical to the strings in kdelibs/kabc/plugins/ldap/resourceldap.cpp,
so you can just copy the translation from that file.

Not included in this patch: Remove the abc resource plugin methods that
implement exactly the same asyncLoad and asyncSave as these default
implementations.

Reviewed by several people.

svn path=/trunk/kdelibs/kabc/; revision=277510