Turn the Organizer of an incidence into a Person object (instead of using just a QString). This has the advantage that it's now possible to have a name and the email specified for the organizer.
So far, the common name setting in the ics file was just discarded, and only the email was used. Now the common name is also read and written to the file.

I don't think I broke free/busy caching (which uses just the email to store the file name on disk), because there I just use the email ( Person::email() ) as before. If f/b doesn't work any more, please tell me...

svn path=/trunk/kdepim/libkcal/; revision=346233