Change the semantics of switching timezones in KOrganizer to ask the user whether she wants to a) keep the events at their absolute (UTC) times, which will make them shift in the calendar view, or b) shift them such that they remain at the same time in the view, which effectively shifts them in UTC. An example, to clarify:
Current timezone is Berlin, my lunch event is at 12:00. I change the
timezone to London. If I chose a) my lunch event is now displayed at
11:00, since that is what 12:00 Berlin time is in London time. This is
the correct thing to do, normally. If I chose b) the event will remain
to to be shown at 12:00, in the view, which means that it has
effectively been shifted in UTC. This is what KOrganizer used to always
do. The reason for this is to make it possible for people to "correct"
their timezone setting, in case they set a timezone in KControl, but not
in KOrganizer (which defaults to UTC, not the KControl timezone), which
would make it necessary to be able to change timezones without shifting
already existing events.

But KOrganizer should always default to the setting KControl, I hear you
scream in agitation. And you are right, I think. We shall make it so.

The texts of the dialogs need some serious polishing, obviously. Ellen 
has volunteered to help with that, so hopefully we'll find texts that 
are not completely confusing.

svn path=/trunk/KDE/kdepim/libkcal/; revision=419477