forward port SVN commit 805701 by vkrause:
Don't remove the observers in the assignment operator, this only makes
sense in the copy ctor where this code was copied from. This fixes
various issues when using the assignment operator, such as wrongly
updated views and most notably the non-persistence of the changes.

ALSO, re-arrange/clean some header includes.

svn path=/trunk/KDE/kdepimlibs/; revision=806192