Adding a helper class for incidence assignment based on the IncidenceBase::Visitor interface.
Using operator=() requires that two Incidence or IncidenceBase pointers are of the same subclass and casted to that one.
The helper resovles the class of one pointer (the one to assign to) through the visitor pattern and then checks via dynamic_cast whether the other one (the assignment's source) matches.


svn path=/trunk/KDE/kdepimlibs/; revision=925461