Introduce the ImapStreamParser as a private class into KIMAP. Build a Session + Job based API on top of it in order to have a neat async API.
Still has a couple of issues (in particular the state() method is
broken). Apart from that the first provided LoginJob just works to
authenticate yourself to an IMAP server.

And sorry for the big drop, I'm no fan of huge commits but this one
grew out of control before I was definitely sold on the fact that the
way I was heading for was the right one.


svn path=/trunk/KDE/kdepimlibs/; revision=943696