Fix KMail always sending mails as base64, even though they could have been sent as 7bit. This was bad because spamassassin gives needlessly base64-encoded text a score of over 2.
The problem is that the mail representation in mimelib and kmime has LF only, no CR.
The LF->CRLF conversion happens in the kio_smtp slave.

Because of that, checking that the number of LF should be the same as the number of CRLF did
not work, since we only have LFs.

Not sure if this workaround is ok.

CCMAIL: exit3219@gmail.com

svn path=/trunk/KDE/kdepimlibs/; revision=1001945