Make receiving a Collection not have the side effect of fetching all Items in it.
When a monitored Collection is added, we will get the Items in it afterward from the Monitor.

When we fetched the Collection ourselves, we do fetch the items.

Except if we're only fetching said Collection becuase it's an ancestor of a Collection we need in the model and isn't there yet.

In that case, we already know we don't need the items in it.

svn path=/trunk/KDE/kdepimlibs/; revision=1010030