kio_nntp could send a broken XOVER request to the server and hang on an error answer. This corrects the broken request.
Thanks to Arkadiusz Miskiewicz for the analysis.

Arkadiusz, I change your patch because the real integer overflow is in the
test "lastSerNum - first > max". Because lastSerNum and first are
"unsigned long", when "lastSerNum" is inferior to "first", the
computation of "lastSerNum - first" overflow and give a big long and as
a consequence the test succeeds.

CCBUG: 218253

svn path=/trunk/KDE/kdepimlibs/; revision=1062490