I admit defeat: Disable the test.
I found no way to fix the test. The problem was that the test manually included ../kmime_headerfactory.cpp, which
was already included by linking to kmime. However, that cpp file had non-exported private symbols, which
is the reason it was included manually again.

Including the same compilation unit twice introduces interesting problems though, such as two instances of 
the same K_GLOBAL_STATIC being created, which caused the tests to fail.

So there is no sane way I found to test private functions. But maybe private functions ought not to be tested.

Anybody an idea how to make the test work again?

svn path=/trunk/KDE/kdepimlibs/; revision=1066358