Add holiday data file for Iran in both Farsi and English.
This is the first holiday file to use alternative calendar systems, both the 
Jalali and Hijri.  Yay!

Note that the days defined in the Hijri calendar can be out by 1 day due to our
using the civil version of the calendar, not the correct lunar version, however
the issue is well know and I hope it isn't an issue with our users.  Hopefully
we will soon get a correct lunar version.

I've copied this wholesale from Wikipedia, so if the Farsi is wrong don't blame
me, feel free to correct :-)

svn path=/trunk/KDE/kdepimlibs/; revision=1131723