Add support for holiday duration.
The existing behaviour of KHolidays for multiday holidays is to return a
Holiday instance for every day the holiday occurs, but this prevents clients
from drawing a single object to cover all the days.  

This change introduces a new mode where multiday holidays can be returned as a
single Holiday instance with a duration value for the number of days the
holiday lasts for.

For backwards compatability, by default multiple Holiday instances will be
still returned but with a duration of 1 day.

New API:
  Holiday::MultidayMode enum to set the multiday handling mode
  Holiday::duration() returns the holiday duration in days
  Holiday::observedStartDate() returns the start date of the holiday
  Holiday::observedEndDate() returns the end date of the holiday
  HolidayRegion::holidays() versions to take the new MultidayMode enum

So an app will now call something like:

  Holiday::List holidays = m_holidayRegion.holidays(startDate, endDate,
                                       Holiday::MultidayHolidaysAsSingleEvents)

Note that the api is subject to revision in light of further enhancements,
and the new enum values might be a bit too long :-)  Comments welcomed.

CCBUG: 187116

svn path=/trunk/KDE/kdepimlibs/; revision=1176150