KHolidays add api to obtain default holiday region code.
Using either the global locale country and language, or a given country and
language, try to find a good default holiday region code.

This code was originally in the Plasma clock widget in 4.5 but is useful for
all clients to obtain a sensible default.  The main change from plasma is if
no default is found a null string is returned rather than the US code.

svn path=/trunk/KDE/kdepimlibs/; revision=1177172