forward port SVN commit 1185335 by winterz:
use the new CalHelper::incOrganizerOwnsCalendar() to determine if the
attendee statuses should be shown in the eventviewer and the tooltips.

so now you can see the up-to-date attendee statuses if you are looking
at the incidence in the organizer's calendar.  For example, if you
are a secretary looking at the boss' calendar you should see the
up-to-date rsvp status of all the attendees.

TODO: forward port CalHelper::incOrganizerOwnsCalendar and use it here.

svn path=/trunk/KDE/kdepimlibs/; revision=1185551