Fetch collection details by default.
If you forget that, the RID safety checks in ResourceBase will silently
drop all changes, so most resources do it. Those with multiple collections
have no choice either way, and for those with only one it is so cheap that
it doesn't matter (it'll be retrieved once and then kept in the Monitor
cache). So, enable it by default. Problem was discovered by David Jarvie
in the KAlarm dir resource.