Fix loaded product ID not being set after loading a KCalCore::Calendar
The loaded product ID was often not set into a KCalCore::Calendar by
FileStorage::load(). It was not set at all when the saved CalFormat
is used to load the calendar. It is set to the write product ID, not
the loaded product ID, when VCalFormat is used to load the calendar.

REVIEW:102189