Fix ancestor retrieval for new collection inside a non-visible parent.
This shows in e.g. KOrganizer when adding a Google or Kolab resource which
have an empty top-level collection (which has no content mimetype matching
the filter) and the actual calendar folders somewhere below that. When
those show up, we need to add the missing links in the ancestor chain as
well. The logic was mostly there, but broken for the (common) case of ETM
showing the full tree (ETM root collection being Collection::root) rather
than just a subtree.

Reviewed by: Christian Mollekopf, Alex Fiestas, Stephen Kelly