Add a new status to AgentBase called "NotConfigured"
When a new resource is created it is typically not configured (the
resource creation is usually divided into instancing the resource and
then configure it), right now resources are setting the status to Broken
in case some important configuration is missing and that can mislead
apps to announce the resource as broken while in reality it is not
configured.

The exampled that motivated this is kimapresource and KMail. Once a new
imap resource is created using KMail, the app will launch a notification
saying that the imap resource is broken, and offline because the status
will say so. That behavior is ok for imap resources that break for some
reason but it is not right for just created resources that are yet not
configured ( and they are expected to no beein configured).

This commit is not changing the deafult Status which is still Idle.

http://lists.kde.org/?l=kde-pim&m=133524726312376&w=2