Don't leave EntitCacheNodes that don't exist on server marked as pending
Calling ensureCached() on an item that is no longer available
on the Akonadi server will now correctly remove the pending tag
from EntityCacheNode before emitting dataAvailable(). Subsequent
call to ensureCached() will not fail.

This could cause ReplayChange operation on an item that was removed
by a previous operation to get stuck forever, because ensuredCache()
would always return false, but would not attempt to fetch the item
from server (becasue the node was marked as pending), and so the
operation (notification) was never removed from Monitor's queue.