Add support for volatile properties, needed to fix a few bugs.
volatile properties are CustomProperties that are only used in
runtime, they will not be written back to ical as they are not
part of the Incidence payload.

Volatile properties start with X-KDE-VOLATILE

This doesn't add new public API and is safe to commit to 4.11.

The first use case for this is to have X-KDE-VOLATILE-AKONADI-ID
that is set internally by the calendar classes, this is needed
because the old KCalCore::Calendar::CalendarObserver::calendarIncidenceDeleted()
has an Incidence as argument and we have no way to get the corresponding Item id
which is key of various hashes.

CCMAIL: winter@kde.org
CCMAIL: chrigi_1@fastmail.fm