Don't crash whem mixing invitations with batch operations.
Creation of the Akonadi::TransactionSequence must be delayed until
it's first usage, otherwise the presence of nested event loops will
cause the transaction to start immediately.

We only want the transaction to start once we add all the sub-jobs
but because of some invitation dialog/nested loop the transaction
started prematurely.