Don't check whether Search resource is online when syncing collection
The akonadi_search_resource is a virtual resource, which is always
online, but AgentManager does not see virtual resources (they
are not exposed from the server), and when queried, it returns
an invalid AgentInstance (where isOnline == false).

We add a simple check for akonadi_search_resource and always return
true from testAndSetOnlineResource() to avoid showing a dialog that
makes no sense in this context.

Proper fix would be to have AgentManager to know about virtual resources,
expose their isVirtual property and have them always set as online
(because virtual resources cannot really be "offline"), but that would
require some changes on the server too. For now we only have this
one virtual resource, so we can stick with a hardcoded name check.