Fix bug where we would send an e-mail with a forged From:
This case happened when we were not the organizer and then
modified the event. An e-mail would be sent with From: <organizer>,
with *all* participants in CC.

The correct is to not send anything, because the user was already
warned via message boz that the event will become out of sync with
the organizer's.

BUG: 289533
BUG: 318394
FIXED-IN: 4.11.3