Use METADATA style annotations only.
This patch add support for METADATA style entries in
GetMetaDataJob/SetMetaDataJob for both METADATA and ANNOTATEMORE mode.
This means an annotation of /shared/comment gets translated in the
ANNOTATEMORE mode to /comment with attribute "value.shared". This
supports of course only a subset of the possible attribute values, but
all that is used by the kimap resource, and as ANNOTATEMORE is anyways
legacy, all that we have to care about (Everything else is still
supported by using the original API).

This is the foundation for the IMAP resource to expose both shared and
private annotations, which we require for kolab. I intend to expose all
annotations including /shared or /private prefix in the
CollectionAnnotationsAttribute, and adapt users of the
CollectionAnnotationsAttribute accordingly. Patches for kdepim-runtime
and kdepim will follow