Use Q_ASSERT instead of Q_VERIFY in AkonadiTest::checkTestIsIsolated()
Q_VERIFY only returns from current method, which is fine when you
use it in the test method, because it interrupts the test. But when
you use it in a method called from the test method, it only returns
from the method, but does not abort the test itself.

In case of checkTestIsIsolated() it gives you a message, but then
messes up your real Akonadi anyway :-)

Q_ASSERT will abort reliably and we don't have to rewrite all unit
tests.