CollectionPathResolver: add API to specify collection to resolve relatively to
Top-level collections don't have to have a unique name (for instance one can have
a calendar and a mail account both called 'Work'), but CPR is too stupid for that
and will try to resolve the path from the first matching collection it finds,
which can lead to false or wrong results.

The new constructor allows specifying a collection that the CPR should resolve the
path relatively to. This can be used by resources for example, which can set their
root collection (which they usually now) and then resolve relatively to it.