ETM: Fixed subtree algorithm in collectionsFetched.
The implemented algorithm relied on specific listing order. It had two problems:
* the subTreesToInsert were IMO never correctly built (the algorithm only works
if children come first, and then it still inserted a parent-child pair to much).
* collectionsToInsert would be overwritten with a dummy collection from the ancestor chain
unless the actual collection comes last in the list (again, depends on listing order)

With this patch the algorithm becomes order independent.

Although the original patch inserted hidden collections in case it has non-hidden
children, this patch removes this behaviour since isHidden recursively checks all ancestors.
(so this scenario shouldn't be possible).
isHidden is potentially broken though since the HIDDEN attribute is probably not
available in the ancestor chain.

REVIEW: 119628