Optimize CollectionSync algorithm
The weakest point of CollectionSync is in findMatchingLocalNode() method
that needs to traverse the LocalNode tree every time from up down to find the
LocalNode* that matches the given remote Collection, which is very inefficient,
especially in the case of initial synchronization.

This change adds pointer to parent node to each LocalNode*, so that we can just
get list of candidates (list of collections with given RemoteID) and then filter
out invalid candidates based on their parents (i.e. for Collection "/foo" with
parent Collection "/bar", the only suitable candidate is LocalNode with RID "/foo"
(which there can be more of) and parent node with RID "/bar" (which already is
unique). This way we filter out nodes /baz/foo and /foo/foo for instance.

I haven't actually benchmarked the code yet (I don't have access to the testing
infrastructure ATM), but I believe it should be fast enough to make initial
sync of 50 000 Collections finish in a reasonable amount of time.