[CollectionSync] Rewrite the algorithm for syncing collections to be much faster
The new algorithm can sync 50 000 collections in about 2 minutes, which is 4 times
faster then what I ever was able to optimize the original algorithm.

The algorithm stores the \"tree\" structure of collections differently than the old
one: it stores list of collections for each parent in a QHash, and then starts
recursively querying Akonadi for collections, easilly comparing list of descendats
it got from Akonadi with the list in the QHash for the same parent RID chain.