Knut Resource: switch from Observer to ObserverV2 and implement itemMoved()
Isolated unit tests that are testing moving items in Akonadi would get weird
change notifications (MOVE, but the item had different RID, and then ADD of
the original item into the original collection, because Knut did not remove
it).

Surprisingly no Akonadi unit tests seem to hit this problem (probably not
tests are checking other Monitor singals and are only comparing IDs, not
RIDs), but Zanshin guys ran into this with their tests.