Monitor: set the session.
Required to properly deal with referenced collections.

With this patch we set the sessionid on the serverside monitor instead
of monitoring each collection that we referenced.
This requires support from the akonadiserver.

The approach is cleaner and keeps the client code cleaner.
It also removes the problem that the mimetypefilter no longer works
because explicitly monitoring collections overrides the mimetypefilter.