Event editor tags in categories field empty
KSqueezedTextLabel has a problem when it comes to resizing. everytime
the window was resized, the text was cleared out. That's why it is empty
at startup, too. Cause at this point the end size isn't set. As a
workaround, we now use a simple QLabel.
QLabel has the disatvantage, that it resizes to its content, so the
widget grows if the content grows. And so the window. With wordwrap=true
multiple lines are allowed, that fits a little bit better, but all in
all this is not a good solution.
The proper solution is to fix KSqueezedTextLabel.

KOLAB: 4177