tagmodel reflect when it is populated
A execution of ItemFetchJob can take some time, so the model should
reflect, when it is usable aka populated. So the TagWidget for example can update
it's view, when all tags are fetched.
Till now the FetchJob was fast enough, that I didn't saw the problem in
uncompleted tag lists, but I can imagine it could be a problem with a
big pile of tags.