Fix handling of read-only mbox files
Up till now a read-only mbox file could not be loaded, since load()
always tried to open in read-write mode.
Now it first tries to open the file readWrite and if that fails,
it retries readOnly.
Also one can now explicitely tell the class to open the file readOnly
even if it could be written.

BUG: 138445
BUG: 323872
FIXED-IN: 4.14.5
REVIEW: 122255